Module names cleanup;
