+++ apps/clients/tests/test_view.py
@@ -6,8 +6,8 @@ from .factories import ClientFactory
 
 from apps.clients.forms import ClientForm
 from apps.clients.models import Client
-from apps.unittest_utils import assert_response_get, assert_response_post
-from apps.globals import PAGINATION_OBJ_COUNT_PER_PAGE
+from apps.unittest_helpers import assert_response_get, assert_response_post
+from apps.constants import PAGINATION_OBJ_COUNT_PER_PAGE
 from ...users.tests import PASSWORD
 from ...users.tests.factories import CxUserFactory
