+++ apps/clients/views.py
@@ -7,8 +7,8 @@ from django.urls import reverse_lazy
 from apps.clients.filters import ClientFilter
 from apps.clients.forms import ClientForm
 from apps.clients.models import Client
-from apps.globals import PAGINATION_OBJ_COUNT_PER_PAGE
-from apps.views_utils import VIEW_MSG, add_error_messages
+from apps.constants import PAGINATION_OBJ_COUNT_PER_PAGE, VIEW_MSG
+from apps.view_helpers import add_error_messages
 
 
 class ClientListView(LoginRequiredMixin, PermissionRequiredMixin, ListView):
