--- apps/globals.py
@@ -1,9 +1,0 @@
-PRODUCT_SAP_DIGITS = 7
-ORDER_SAP_DIGITS = 8
-CLIENT_SAP_DIGITS = 7
-
-FLOAT_DEFAULT = 1.0
-INT_DEFAULT = 1
-
-PAGINATION_LINKS_MAX_COUNT = 20
-PAGINATION_OBJ_COUNT_PER_PAGE = 10
