+++ apps/orders/forms.py
@@ -4,7 +4,7 @@ from django.forms.models import inlineformset_factory
 
 from bootstrap_datepicker_plus import DatePickerInput
 
-from apps.forms_utils import SAP_STYLE, NUM_STYLE_NO_REQ, BASIC_NO_HINTS_STYLE, INPUT_MEASUREMENT_FORM_STYLE_50px, \
+from apps.form_styles import SAP_STYLE, NUM_STYLE_NO_REQ, BASIC_NO_HINTS_STYLE, INPUT_MEASUREMENT_FORM_STYLE_50px, \
     INPUT_MEASUREMENT_FORM_STYLE_70px, INPUT_MEASUREMENT_FORM_STYLE_71px, ORDER_SAP_STYLE, NUM_STYLE, BASIC_REQ_STYLE, \
     INT_STYLE, PALLET_NUMBER_STYLE
 from apps.orders.models import Order, MeasurementReport, Measurement
