+++ apps/products/views.py
@@ -9,8 +9,8 @@ from django.views.generic import UpdateView, CreateView, DetailView, DeleteView,
 from apps.products.filters import ProductFilter
 from apps.products.forms import ProductForm, SpecificationForm, ProductSpecificationMultiForm
 from apps.products.models import Product
-from apps.globals import PAGINATION_OBJ_COUNT_PER_PAGE
-from apps.views_utils import VIEW_MSG, add_error_messages
+from apps.constants import PAGINATION_OBJ_COUNT_PER_PAGE, VIEW_MSG
+from apps.view_helpers import add_error_messages
 
 
 class ProductListView(LoginRequiredMixin, PermissionRequiredMixin, ListView):
